requirements.txt
flask
flask-cors
ultralytics
deepface
pillow
opencv-python-headless
numpy
